Select plain or redux version in app.js
Introduced a bool to change between plain react and react-redux version
of the components. At the moment it's just a hard-coded bool but could
make this something switchable in the UI.

Also does the first store dispatch of the UserAuthenticated action.
There's room here to look at stronger typing of the store now.