Login container & component
Provide a means of "logging in" if the auth implementation states that
the user isn't currently logged in. At the moment this is just a button
that changes the auth implementation isAuth... setting but this will
probably be what initiates the redirect to Spotify OAuth page

Doesn't render anything if the user is already logged in.