Undo most of the styling
I wasn't happy with the way the styling was going so want to go for
something else. This is getting rid of what was there.

Also removes the UI reducer and action creators. These may come back
later but the menu show/hide isn't needed now so delete it all for now.