Really naive save button behaviour
Disable the button when a new playlist is selected, enable it when a new
track is added.

I think I'll probably track changes to the playlist in the future so
instead of a bool "canSave" it'll look at whether there are any pending
changes to the playlist. On save or undo of all changes the button can
act accordingly with regards to being enabled.