Mark a track as either persisted or not
In order to determine which tracks of the playlist should be sent in the
save tracks HTTP request, the app needs to track whether each track
either already present (i.e. had been saved to the playlist prior) or
was added to the playlist via this app.

With a simple "isPersisted" bool, the app can filter which track URIs to
send in the HTTP request